MIPT_MLOps_HW_2/requirements.txt
torch>=2.0.0
pytorch-lightning>=2.0.0
hydra-core>=1.3.0
dvc[s3]>=3.0.0
pybind11>=2.11.0
pandas
pytest